refactor: move CLI command files into src/cli/ (#260)
Move the nine one-shot CLI command modules into src/cli/:

  init, status, install, backup, import, daemon,
  web-daemon, proxy-daemon, tui

Stay at src/ root:
  - index.ts (binary entrypoint — package.json bin target)
  - app.ts, web.ts, proxy.ts (long-running server processes)
  - all runtime/shared modules (runtime, context, session, memory,
    registry, model, config, log, util, server, pairing, kernel,
    segments, queue, global-config)

Pure rename + import path rewrite. No behavior change, no
changelog entry. Build clean, smoke tests (`kern --help`,
`kern status`) pass.
